requirements.txt
ansio==0.0.1
beautifulsoup4==4.12.3
docker==7.1.0
duckduckgo-search==6.1.12
faiss-cpu==1.8.0.post1
flask[async]==3.0.3
flask-basicauth==0.2.0
inputimeout==1.0.4
langchain-anthropic==0.1.19
langchain-community==0.2.7
langchain-google-genai==1.0.7
langchain-groq==0.1.6
langchain-huggingface==0.0.3
langchain-mistralai==0.1.8
langchain-ollama==0.1.3
langchain-openai==0.1.15
lxml_html_clean==0.3.1
markdown==3.7
newspaper3k==0.2.8
paramiko==3.4.0
pypdf==4.3.1
python-dotenv==1.0.1
requests-oauthlib==1.3.1
sentence-transformers==3.0.1
unstructured==0.15.13
unstructured-client==0.25.9
webcolors==24.6.0
